+++ scripts/lumitools/gen_CPE_VisibilityHookData.py
@@ -4,22 +4,31 @@
 # Input arguments
 #   * CPEpackages_dir : Directory with the CPE definitinon files (in .csv format)
 #   * VisibilityHookData_dir : Directory with the LMOD modulerc files
-#   * version : Release of the CPE to generate the file for.
+#   * version_stack : Release of the CPE to generate the file for (and providing the list of modules)
+#   * version_alias: Release of the CPE to use for the module versions
 #
 
-def gen_CPE_VisibilityHookData( CPEpackages_dir, VisibilityHookData_dir, version ):
+import re
 
-    def write_package( fileH, PEpackage, module, package_versions ):
+def gen_CPE_VisibilityHookData( CPEpackages_dir, VisibilityHookData_dir, version_stack, version_alias ):
+
+    def write_package( fileH, PEpackage, module, package_versions, minv='00.00', maxv='99.99' ):
 
         # Note that if a particular PEpackage does not exist in package_versions, no
         # error is printed. This is done in this way to be able to cope with evolutions
         # in the packages while still having a single script that works for all as
         # those packages will now simply be skipped.
 
-        if PEpackage in package_versions:
-            version = package_versions[PEpackage]
-            fileH.write( "['%s']='%s'," % (module, version) )
-
+        nonlocal version_stack
+        nonlocal version_alias
+        
+        nversion_stack = re.sub( '\D+', '', version_stack )
+        nversion_alias = re.sub( '\D+', '', version_alias )
+        nminv =          re.sub( '\D+', '', minv )
+        nmaxv =          re.sub( '\D+', '', maxv )
+        if nversion_stack >= nminv and nversion_stack <= nmaxv and nversion_alias >= nminv and nversion_alias <= nmaxv and PEpackage in package_versions:
+            mversion = package_versions[PEpackage]
+            fileH.write( f"['{module}']='{mversion}'," )
 
     #
     # Core of the gen_EB_external_modules_from_CPEdef function
@@ -31,12 +40,12 @@ def gen_CPE_VisibilityHookData( CPEpackages_dir, VisibilityHookData_dir, version
     #
     # Read the .csv file with toolchain data.
     #
-    CPEpackages_file = os.path.join( CPEpackages_dir, 'CPEpackages_' + version + '.csv' )
-    print( 'Reading the toolchain composition from %s.' % CPEpackages_file )
+    CPEpackages_file = os.path.join( CPEpackages_dir, f'CPEpackages_{version_stack}.csv' )
+    print( f'Reading the toolchain composition from {CPEpackages_file}.' )
     try:
         fileH = open( CPEpackages_file, 'r' )
     except OSerror:
-        print( 'Failed to open the toolchain packages file %s.' % CPEpackages_file )
+        print( f'Failed to open the toolchain packages file {CPEpackages_file}.' )
         exit()
 
     package_versions = {}
@@ -49,18 +58,58 @@ def gen_CPE_VisibilityHookData( CPEpackages_dir, VisibilityHookData_dir, version
         package_versions[row[0]] = row[1]
 
     fileH.close()
+    
+    #
+    # If version_alias is different from version_stack: Also read the CPEpackages file
+    # for version_alias and correct package_versions with the version information from
+    # the version_alias CPEpackages file.
+    #
+    if version_stack != version_alias :
+        
+        CPEpackages_file = os.path.join( CPEpackages_dir, f'CPEpackages_{version_alias}.csv' )
+        print( f'Reading the toolchain composition for the alias versions from {CPEpackages_file}.' )
+        try:
+            fileH = open( CPEpackages_file, 'r' )
+        except OSerror:
+            print( f'Failed to open the toolchain packages file {CPEpackages_file}.' )
+            exit()
+    
+        package_versions_alias = {}
+    
+        package_reader = csv.reader( fileH )
+        # Skip the header line
+        next( package_reader )
+        # Read the data and build the package_versions dictionary
+        for row in package_reader :
+            package_versions_alias[row[0]] = row[1]
+    
+        fileH.close()
+        
+        #
+        # Now update the package versions with the aliased ones
+        #
+        
+        to_delete = []
+        for key in package_versions :
+            if key in package_versions_alias :
+                package_versions[key] = package_versions_alias[key]
+            else :
+                to_delete.append( key )
+                
+        for key in to_delete:
+            del( package_versions[key] )
 
     #
     # Add missing packages or entries needed for this script
     #
-    package_versions['CPE'] = version
+    package_versions['CPE'] = version_stack
 
 
     #
     # Open the CPE-specific modulerc file
     #
     print( 'Installing in: %s' % VisibilityHookData_dir )
-    extdeffile = 'CPEmodules_%s.lua' % version.replace( '.', '_' )
+    extdeffile = 'CPEmodules_%s.lua' % version_stack.replace( '.', '_' )
     extdeffileanddir = os.path.join( VisibilityHookData_dir, extdeffile )
     print( 'Generating %s...' % extdeffileanddir )
     fileH = open( extdeffileanddir, 'w' )
@@ -78,11 +127,22 @@ def gen_CPE_VisibilityHookData( CPEpackages_dir, VisibilityHookData_dir, version
     write_package( fileH, 'cpe-prgenv',           'PrgEnv-aocc',              package_versions )
     write_package( fileH, 'cpe-prgenv',           'PrgEnv-intel',             package_versions )
     write_package( fileH, 'cpe-prgenv',           'PrgEnv-nvidia',            package_versions )
+    write_package( fileH, 'cpe-prgenv',           'PrgEnv-nvhpc',             package_versions, minv='22.06' )
+    write_package( fileH, 'cpe-prgenv',           'PrgEnv-gnu-amd',           package_versions, minv='22.08' )
+    write_package( fileH, 'cpe-prgenv',           'PrgEnv-cray-amd',          package_versions, minv='22.08' )
 
     write_package( fileH, 'CCE',                  'cce',                      package_versions )
     write_package( fileH, 'GCC',                  'gcc',                      package_versions )
     write_package( fileH, 'AOCC',                 'aocc',                     package_versions )
+    write_package( fileH, 'ROCM',                 'amd',                      package_versions )
     write_package( fileH, 'intel',                'intel',                    package_versions )
+
+    write_package( fileH, 'CCE',                  'cce-mixed',                package_versions, minv='22.06' )
+    write_package( fileH, 'GCC',                  'gcc_mixed',                package_versions, minv='22.06' )
+    write_package( fileH, 'AOCC',                 'aocc_mixed',               package_versions, minv='22.06' )
+    write_package( fileH, 'ROCM',                 'amd_mixed',                package_versions, minv='22.06' )
+
+    write_package( fileH, 'ROCM',                 'rocm',                     package_versions )
 
     write_package( fileH, 'craype',               'craype',                   package_versions )
     write_package( fileH, 'CPE',                  'cpe',                      package_versions )
@@ -94,8 +154,9 @@ def gen_CPE_VisibilityHookData( CPEpackages_dir, VisibilityHookData_dir, version
     write_package( fileH, 'MPICH',                'cray-mpich',               package_versions )
     write_package( fileH, 'MPICH',                'cray-mpich-abi',           package_versions )
     write_package( fileH, 'PMI',                  'cray-pmi',                 package_versions )
-    write_package( fileH, 'PMI',                  'cray-pmi-lib',             package_versions )
+    write_package( fileH, 'PMI',                  'cray-pmi-lib',             package_versions, maxv='22.06' )
     write_package( fileH, 'OpenSHMEMX',           'cray-openshmemx',          package_versions )
+    write_package( fileH, 'MPIxlate',             'cray-mpixlate',            package_versions, minv='21.11' )
 
     write_package( fileH, 'FFTW',                 'cray-fftw',                package_versions )
     write_package( fileH, 'HDF5',                 'cray-hdf5',                package_versions )
@@ -110,6 +171,7 @@ def gen_CPE_VisibilityHookData( CPEpackages_dir, VisibilityHookData_dir, version
     write_package( fileH, 'CCDB',                 'cray-ccdb',                package_versions )
     write_package( fileH, 'CTI',                  'cray-cti',                 package_versions )
     write_package( fileH, 'DSMML',                'cray-dsmml',               package_versions )
+    write_package( fileH, 'cray-dyninst',         'cray-dyninst',             package_versions, minv='21.12' )
     write_package( fileH, 'jemalloc',             'cray-jemalloc',            package_versions )
     write_package( fileH, 'STAT',                 'cray-stat',                package_versions )
     write_package( fileH, 'craypkg-gen',          'craypkg-gen',              package_versions )